added authorization checks for editing and deleting posts to work only for the owner/creator

+++ controllers/feed.js
@@ -134,6 +134,11 @@ exports.updatePost = (req, res, next) => {
         error.statuscode = 404; // 404 - not found code
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!')
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -162,9 +167,14 @@ exports.deletePost = (req, res, next) => {
         error.statuscode = 404; // 404 - not found code
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!')
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user is the creator
       clearImage(post.imageUrl);
-      return Post.findByIdAndRemove(postId)
+      return Post.findByIdAndRemove(postId);
     })
     .then(result => {
       console.log(result);
